refactor(migrate-central): extract migrate_import module (OI-1537/1539, part 1/3)
Move _import_table (OI-1537, ~153 lines) + _compare_counts (OI-1539, ~125 lines)
and their direct helpers into new scripts/lib/migrate_import.py.

Extracted to migrate_import.py:
- _import_table, _compare_counts, _collect_skipped_rows
- _record_skip, _record_rowid_mapping, _mapped_central_rowid, _resolve_prior_skip
- _src_table_present
- _is_collision_column, _collect_collision_columns, _collect_json_array_columns
- _prefix_value, _prefix_json_array
- _table_exists, _table_columns, _column_exists, _integer_primary_key, _common_columns
- ImportSummary dataclass, _now_utc_iso
- COLLISION_* constants, _IMPORT_BATCH_SIZE, ABORT_FLAG, check_abort, AbortRequested

migrate_to_central_vnx.py re-imports all moved names for backward compatibility.
Existing callers (migrate_dry_run.py, tests via M.*) unchanged.

tests/test_migrate_to_central_vnx.py: fixture patches _MI.ABORT_FLAG alongside
M.ABORT_FLAG so check_abort() in migrate_import sees the test-overridden path.

Dispatch-ID: 20260526-oi-b1a-migrate-import
